End To End Analyzer App From implementing To deployment/pyproject.toml
[project]
name = "cv1"
version = "0.1.0"
requires-python = ">=3.14.2"
dependencies = [
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.0",
    "python-dotenv>=1.2.1",
]
